refactor: pages > mystore > create > index.tsx 리펙토링
