+++ src/widgets/createStore/index.ts
@@ -1,1 +1,2 @@
 export { InputContent } from '@/widgets/createStore/ui/inputContent/InputContent';
+export { ModalContainer } from './ui/modalContainer/ModalContainer';
