+++ src/widgets/mystore/ui/StoreContent/StoreContent.styled.ts
@@ -1,0 +1,9 @@
+
+
+const StoreContent.styled = () => {
+  return (
+    <div>StoreContent.styled</div>
+  )
+}
+
+export default StoreContent.styled
